feat(backend): Move Settings to AutoGPTLib & Spin-up Database manager on rest.py
