+++ autogpt_platform/autogpt_libs/autogpt_libs/supabase_integration_credentials_store/__init__.py
@@ -1,5 +1,5 @@
 from .store import SupabaseIntegrationCredentialsStore
-from .types import Credentials, APIKeyCredentials, OAuth2Credentials
+from .types import APIKeyCredentials, Credentials, OAuth2Credentials
 
 __all__ = [
     "SupabaseIntegrationCredentialsStore",
