+++ autogpt_platform/backend/backend/blocks/google/_auth.py
@@ -1,10 +1,10 @@
 from typing import Literal
 
 from autogpt_libs.supabase_integration_credentials_store.types import OAuth2Credentials
+from autogpt_libs.utils.settings import Secrets
 from pydantic import SecretStr
 
 from backend.data.model import CredentialsField, CredentialsMetaInput
-from backend.util.settings import Secrets
 
 # --8<-- [start:GoogleOAuthIsConfigured]
 secrets = Secrets()
