+++ autogpt_platform/backend/backend/data/credit.py
@@ -1,6 +1,7 @@
 from abc import ABC, abstractmethod
 from datetime import datetime, timezone
 
+from autogpt_libs.utils.settings import Config
 from prisma import Json
 from prisma.enums import UserBlockCreditType
 from prisma.errors import UniqueViolationError
@@ -9,7 +10,6 @@ from prisma.models import UserBlockCredit
 from backend.data.block import Block, BlockInput, get_block
 from backend.data.block_cost_config import BLOCK_COSTS
 from backend.data.cost import BlockCost, BlockCostType
-from backend.util.settings import Config
 
 config = Config()
