+++ autogpt_platform/backend/backend/executor/manager.py
@@ -17,6 +17,7 @@ if TYPE_CHECKING:
     from backend.executor import DatabaseManager
 
 from autogpt_libs.utils.cache import thread_cached
+from autogpt_libs.utils.settings import Settings
 
 from backend.blocks.agent import AgentExecutorBlock
 from backend.data import redis
@@ -43,7 +44,6 @@ from backend.util.service import (
     expose,
     get_service_client,
 )
-from backend.util.settings import Settings
 from backend.util.type import convert
 
 logger = logging.getLogger(__name__)
