+++ autogpt_platform/backend/backend/integrations/creds_manager.py
@@ -6,12 +6,12 @@ from autogpt_libs.supabase_integration_credentials_store import (
     Credentials,
     SupabaseIntegrationCredentialsStore,
 )
+from autogpt_libs.utils.settings import Settings
 from autogpt_libs.utils.synchronize import RedisKeyedMutex
 from redis.lock import Lock as RedisLock
 
 from backend.data import redis
 from backend.integrations.oauth import HANDLERS_BY_NAME, BaseOAuthHandler
-from backend.util.settings import Settings
 
 logger = logging.getLogger(__name__)
 settings = Settings()
@@ -129,7 +129,6 @@ class IntegrationCredentialsManager:
 
     def _acquire_lock(self, user_id: str, credentials_id: str, *args: str) -> RedisLock:
         key = (
-            self.store.db_manager,
             f"user:{user_id}",
             f"credentials:{credentials_id}",
             *args,
