+++ autogpt_platform/backend/backend/rest.py
@@ -1,5 +1,5 @@
 from backend.app import run_processes
-from backend.executor import ExecutionScheduler
+from backend.executor import DatabaseManager, ExecutionScheduler
 from backend.server.rest_api import AgentServer
 
 
@@ -8,6 +8,7 @@ def main():
     Run all the processes required for the AutoGPT-server REST API.
     """
     run_processes(
+        DatabaseManager(),
         ExecutionScheduler(),
         AgentServer(),
     )
