+++ autogpt_platform/backend/backend/server/integrations/utils.py
@@ -1,6 +1,5 @@
+from autogpt_libs.utils.settings import Settings
 from supabase import Client, create_client
-
-from backend.util.settings import Settings
 
 settings = Settings()
