+++ autogpt_platform/backend/backend/server/rest_api.py
@@ -2,6 +2,7 @@ import contextlib
 import logging
 import typing
 
+import autogpt_libs.utils.settings
 import fastapi
 import fastapi.responses
 import starlette.middleware.cors
@@ -13,9 +14,8 @@ import backend.data.graph
 import backend.data.user
 import backend.server.routers.v1
 import backend.util.service
-import backend.util.settings
 
-settings = backend.util.settings.Settings()
+settings = autogpt_libs.utils.settings.Settings()
 logger = logging.getLogger(__name__)
 
 
@@ -46,7 +46,7 @@ def handle_internal_http_error(status_code: int = 500, log_error: bool = True):
 
 docs_url = (
     "/docs"
-    if settings.config.app_env == backend.util.settings.AppEnvironment.LOCAL
+    if settings.config.app_env == autogpt_libs.utils.settings.AppEnvironment.LOCAL
     else None
 )
 
@@ -83,8 +83,8 @@ class AgentServer(backend.util.service.AppProcess):
         )
         uvicorn.run(
             server_app,
-            host=backend.util.settings.Config().agent_api_host,
-            port=backend.util.settings.Config().agent_api_port,
+            host=autogpt_libs.utils.settings.Config().agent_api_host,
+            port=autogpt_libs.utils.settings.Config().agent_api_port,
         )
 
     @staticmethod
