+++ autogpt_platform/backend/backend/server/utils.py
@@ -1,8 +1,7 @@
 from autogpt_libs.auth.depends import requires_user
 from autogpt_libs.auth.models import User
+from autogpt_libs.utils.settings import Settings
 from fastapi import Depends
-
-from backend.util.settings import Settings
 
 settings = Settings()
