+++ autogpt_platform/backend/backend/util/request.py
@@ -4,8 +4,7 @@ from typing import Callable
 from urllib.parse import urlparse
 
 import requests as req
-
-from backend.util.settings import Config
+from autogpt_libs.utils.settings import Config
 
 # List of IP networks to block
 BLOCKED_IP_NETWORKS = [
